챗봇 api에 파일 전송 수정2

+++ cym-back/src/main/java/CY/cymake/Domain/Chat/ChatService.java
@@ -32,8 +32,11 @@ public class ChatService {
     private final S3Service s3Service;
     @Transactional
     public ChatResDto sendToFlask(String sessionId, String question, String companyCode, MultipartFile file) throws IOException {
-        String fileUrl = s3Service.uploadFile(file, "files/search/" + file.getOriginalFilename());
-        System.out.println(fileUrl);
+        String fileUrl = null;
+        if(file != null) {
+            fileUrl = s3Service.uploadFile(file, "files/search/" + file.getOriginalFilename());
+            System.out.println(fileUrl);
+        }
         ChatReqDto chatReqDto = new ChatReqDto(companyCode, question, sessionId, fileUrl);
 
         RestTemplate restTemplate = new RestTemplate();
@@ -52,8 +55,10 @@ public class ChatService {
 
         //Flask 서버로 데이터를 전송하고 받은 응답 값을 return
         String result = restTemplate.postForObject(url, entity, String.class);
-        //파일 s3에서 삭제
-        s3Service.deleteFile("files/search/", file.getOriginalFilename());
+        if(fileUrl != null) {
+            //파일 s3에서 삭제
+            s3Service.deleteFile("files/search/", file.getOriginalFilename());
+        }
         // String을 JSON 형식으로 변환
         Map<String, Object> responseMap = new HashMap<>();
         // String을 JSON 형식으로 변환
